requirements.txt
flask==3.0.3
flask-cors==4.0.1
tensorflow==2.15.0
numpy==1.26.4
pillow==10.3.0
gunicorn==22.0.0
